Move WrapperU256 type into a common-types library
