+++ umbrella-feeds-reader/Cargo.toml
@@ -9,10 +9,9 @@ authors = [ "Concordium <developers@concordium.com>" ]
 description = "An oracle system"
 
 [features]
-default = ["std", "u256_amount"]
+default = ["std"]
 std = ["concordium-std/std"]
 wee_alloc = ["concordium-std/wee_alloc"]
-u256_amount = []
 
 [dev-dependencies]
 concordium-smart-contract-testing = "3.0"
@@ -27,9 +26,8 @@ path = "../staking-bank/"
 [dependencies.umbrella-feeds]
 path = "../umbrella-feeds/"
 
-[dependencies.primitive-types]
-version = "0.11"
-default-features = false
+[dependencies.common-types]
+path = "../common-types/"
 
 [dependencies]
 concordium-std = {version = "8.0", default-features = false}
